+++ pkgs/test_api/lib/src/expect/future_matchers.dart
@@ -3,11 +3,11 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:matcher/matcher.dart';
+import 'package:test_api/test_api.dart' show pumpEventQueue;
 
-import '../util/pretty_print.dart';
 import 'async_matcher.dart';
 import 'expect.dart';
-import 'utils.dart';
+import 'util/pretty_print.dart';
 
 /// Matches a [Future] that completes successfully with any value.
 ///