+++ pkgs/test_api/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.4.0-dev
 
+* Add libraries `scaffolding.dart`, and `expect.dart` to allow importing as
+  subset of the normal surface area.
+* Add new APIs in `hooks.dart` to allow writing custom expectation frameworks
+  which integrate with the test runner.
 * Add examples to `throwsA` and make top-level `throws...` matchers refer to it.
 * Disable stack trace chaining by default.
 * Fix `expectAsync` function type checks.